requirements.txt
pickle
sklearn
streamlit
lightgbm
xgboost
pandas
numpy
matplotlib.pyplot
seaborn
imblearn.over_sampling
warnings